+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Marking a lot of articles as read at once now shows a confirmation dialog to prevent accidental actions. (#560)
+- Thumbnail images in article list will now display in compact mode if enabled. (#589)
 
 ### Fixed